requirements.txt
fastapi
google-cloud-aiplatform
python-dotenv